backend/requirements.txt
Flask
Flask-Cors
google-genai
python-dotenv